+++ Tests/FoundationTests/NSAttributedStringExtensionsTests.swift
@@ -97,29 +97,34 @@ final class NSAttributedStringExtensionsTests: XCTestCase {
         let attrAtBeginning = attrTestString.attributes(at: 0, effectiveRange: nil)
         XCTAssert(attrAtBeginning.count == 1)
         
+        var passed = false
         // iterate through each range of attributes
         attrTestString.enumerateAttributes(in: NSRange(0..<attrTestString.length), options: .longestEffectiveRangeNotRequired) { attrs, range, _ in
             
             let emailFromRange = attrTestString.attributedSubstring(from: range).string
             
-            // exit if
-            guard attrs.count > attrAtBeginning.count else {
-                XCTAssertNotEqual(emailFromRange, email)
-                return
-            }
+            // exit if there are not more attributes for the subsequence than what was there originally
+            guard attrs.count > attrAtBeginning.count else { return }
 
+            // confirm that the string with the applied attributes is the email
             XCTAssertEqual(emailFromRange, email)
+            
+            // the range contains the email, check to make sure the attributes are there and correct
             for attr in attrs {
                 if attr.key == .underlineStyle {
                     XCTAssertEqual(attr.value as? NSUnderlineStyle.RawValue, NSUnderlineStyle.styleSingle.rawValue)
+                    passed = true
                 } else if attr.key == .foregroundColor {
                     XCTAssertEqual(attr.value as? UIColor, UIColor.blue)
+                    passed = true
                 } else if attr.key == .font {
                     XCTAssertEqual((attr.value as? UIFont), .boldSystemFont(ofSize: UIFont.systemFontSize))
                 } else {
-                    XCTAssert(false)
+                    passed = false
                 }
             }
+            
+            XCTAssert(passed)
         }
     }
     
@@ -129,27 +134,37 @@ final class NSAttributedStringExtensionsTests: XCTestCase {
         let attrGreeting = NSAttributedString(string: greeting).italicized.applying(
             attributes: [.underlineStyle: NSUnderlineStyle.styleSingle.rawValue,
                          .foregroundColor: UIColor.red], toOccurrencesOf: name)
+        
         let attrAtBeginning = attrGreeting.attributes(at: 0, effectiveRange: nil)
+        // assert that there is only one attribute at beginning from italics
         XCTAssertEqual(attrAtBeginning.count, 1)
         
+        var passed = false
+        // iterate through each range of attributes
         attrGreeting.enumerateAttributes(in: NSRange(0..<attrGreeting.length), options: .longestEffectiveRangeNotRequired) { attrs, range, _ in
+            // exit if there are not more attributes for the subsequence than what was there originally
             guard attrs.count > attrAtBeginning.count else { return }
             
+            // confirm that the attributed string is the name
             let stringAtRange = attrGreeting.attributedSubstring(from: range).string
             XCTAssertEqual(stringAtRange, name)
             
             for attr in attrs {
                 if attr.key == .underlineStyle {
                     XCTAssertEqual(attr.value as? NSUnderlineStyle.RawValue, NSUnderlineStyle.styleSingle.rawValue)
+                    passed = true
                 } else if attr.key == .foregroundColor {
                     XCTAssertEqual(attr.value as? UIColor, UIColor.red)
+                    passed = true
                 } else if attr.key == .font {
                     XCTAssertEqual((attr.value as? UIFont), .italicSystemFont(ofSize: UIFont.systemFontSize))
                 } else {
-                    XCTAssert(false)
+                    passed = false
                 }
             }
         }
+        
+        XCTAssert(passed)
     }
     #endif